\name{danish}
\alias{danish}
\title{Danish Fire Insurance Claims} 
\description{
  These data describe large fire insurance claims 
  in Denmark from Thursday 3rd January 1980 until
  Monday 31st December 1990.
  The data are contained in a numeric vector.
  The dates of each observation are contained in a \code{times}
  attribute, which is an object of class \code{"POSIXct"} (see
  \code{\link{DateTimeClasses}}).
  They were supplied by Mette Rytgaard of Copenhagen Re. 
  Note that these data form an irregular time series. 
}
\usage{data(danish)}
\format{A numeric vector containing 2167 observations, with a
  \code{times} attribute which is a \code{POSIXct} object
  of the same length.}
\keyword{datasets}
