\name{nidd.annual}
\alias{nidd.annual}
\title{The River Nidd Data} 
\description{
  These data represent annual maximal levels of the 
  River Nidd in Yorkshire. 
  These data are suitable for analysis with \code{gev}. 
}
\format{A numeric vector containing 35 observations.}
\usage{data(nidd.annual)}
\keyword{datasets}
