\name{nidd.thresh}
\alias{nidd.thresh}
\title{The River Nidd Data} 
\description{
  These data represent high river levels of the 
  River Nidd in Yorkshire above a threshold value 
  of 65. These data are suitable for analysis with 
  \code{gpd}.  
}
\format{A numeric vector containing 154 observations.}
\usage{data(nidd.thresh)}
\keyword{datasets}
