\name{sp.raw}
\alias{sp.raw}
\title{SP Data to June 1993} 
\description{
  The daily closing values of the S&P index from
  Monday 4th January 1960 until Friday 11th June 1993.
  The data are contained in a numeric vector.
  The dates of each observation are contained in a \code{times}
  attribute, which is an object of class \code{"POSIXct"} (see
  \code{\link{DateTimeClasses}}).
}
\format{A numeric vector containing 8415 observations, with a
  \code{times} attribute which is a \code{POSIXct} object
  of the same length.}
\usage{data(sp.raw)}
\keyword{datasets}
