\name{spto87}
\alias{spto87}
\title{SP Return Data to October 1987} 
\description{
  The daily log returns on the S&P index value from
  Tuesday 5th January 1960 until Friday 16 October 1987.
  The data are contained in a numeric vector.
  The dates of each observation are contained in a \code{times}
  attribute, which is an object of class \code{"POSIXct"} (see
  \code{\link{DateTimeClasses}}).
}
\format{A numeric vector containing 6985 observations, with a
  \code{times} attribute which is a \code{POSIXct} object
  of the same length.}
\usage{data(spto87)}
\keyword{datasets}
